+++ src/promptflow-devkit/promptflow/_proxy/_base_executor_proxy.py
@@ -13,8 +13,8 @@ import httpx
 
 from promptflow._constants import DEFAULT_ENCODING, LINE_TIMEOUT_SEC
 from promptflow._core._errors import NotSupported, UnexpectedError
+from promptflow._proxy._errors import ExecutorServiceUnhealthy
 from promptflow._sdk._constants import (
-    DAG_FILE_NAME,
     FLOW_META_JSON,
     FLOW_META_JSON_GEN_TIMEOUT,
     FLOW_TOOLS_JSON,
@@ -23,10 +23,9 @@ from promptflow._sdk._constants import (
 )
 from promptflow._utils.async_utils import async_run_allowing_running_loop
 from promptflow._utils.exception_utils import ErrorResponse, ExceptionPresenter
-from promptflow._utils.flow_utils import is_flex_flow, read_json_content
+from promptflow._utils.flow_utils import is_flex_flow, read_json_content, resolve_flow_path
 from promptflow._utils.logger_utils import bulk_logger
 from promptflow._utils.utils import load_json
-from promptflow.batch._errors import ExecutorServiceUnhealthy
 from promptflow.contracts.run_info import FlowRunInfo
 from promptflow.exceptions import ErrorTarget, ValidationException
 from promptflow.executor._errors import AggregationNodeExecutionTimeoutError, LineExecutionTimeoutError
@@ -37,6 +36,31 @@ EXECUTOR_UNHEALTHY_MESSAGE = "The executor service is currently not in a healthy
 
 
 class AbstractExecutorProxy:
+
+    def __init__(self):
+        self._should_apply_inputs_mapping = True
+        self._allow_aggregation = True
+
+    @property
+    def should_apply_inputs_mapping(self):
+        """should apply inputs mapping when process batch inputs.
+        For normal batch runs, proxy should apply column mapping right after process batch inputs
+        For chat group batch runs, proxy only resolve inputs to a list
+        and inputs mapping will be applied in orchestrator
+        :return: _description_
+        :rtype: _type_
+        """
+        return self._should_apply_inputs_mapping
+
+    @property
+    def allow_aggregation(self):
+        """whether allow to run aggregation.
+        Chat group batch runs do not support aggregation
+        :return: _description_
+        :rtype: _type_
+        """
+        return self._allow_aggregation
+
     @classmethod
     def dump_metadata(cls, flow_file: Path, working_dir: Path) -> NoReturn:
         """Generate metadata for a specific flow."""
@@ -53,7 +77,7 @@ class AbstractExecutorProxy:
         load_in_subprocess: bool = True,
     ) -> dict:
         """Generate flow.tools.json for the specified flow."""
-        if is_flex_flow(file_path=flow_file, working_dir=working_dir):
+        if is_flex_flow(flow_path=flow_file, working_dir=working_dir):
             return {}
         else:
             return cls._generate_flow_tools_json(flow_file, working_dir, dump, timeout, load_in_subprocess)
@@ -94,7 +118,7 @@ class AbstractExecutorProxy:
         :return: The metadata of the flow.
         :rtype: Dict[str, Any]
         """
-        if is_flex_flow(file_path=flow_file, working_dir=working_dir):
+        if is_flex_flow(flow_path=flow_file, working_dir=working_dir):
             return cls._generate_flow_json(flow_file, working_dir, dump, timeout, load_in_subprocess)
         else:
             return {}
@@ -128,6 +152,14 @@ class AbstractExecutorProxy:
         connections: Optional[dict] = None,
         storage: Optional[AbstractRunStorage] = None,
         init_kwargs: Optional[Dict[str, Any]] = None,
+        # below parameters are added for multi-container
+        # executor_client is provided by runtime PythonExecutorClient class
+        executor_client: Optional[Any] = None,
+        environment_variables: dict = None,
+        log_path: Optional[Path] = None,
+        output_dir: Optional[Path] = None,
+        worker_count: Optional[int] = None,
+        line_timeout_sec: Optional[int] = None,
         **kwargs,
     ) -> "AbstractExecutorProxy":
         """Create a new executor"""
@@ -178,6 +210,7 @@ class APIBasedExecutorProxy(AbstractExecutorProxy):
                 where we can find metadata under .promptflow. Will use current working directory if not provided.
         :type working_dir: Path
         """
+        super().__init__()
         self._working_dir = working_dir or Path.cwd()
         self._enable_stream_output = enable_stream_output
 
@@ -240,8 +273,9 @@ class APIBasedExecutorProxy(AbstractExecutorProxy):
         """Get the inputs definition of an eager flow"""
         from promptflow.contracts.flow import FlowInputDefinition
 
+        _, flow_file = resolve_flow_path(flow_path=self.working_dir, check_flow_exist=False)
         flow_meta = self.generate_flow_json(
-            flow_file=self.working_dir / DAG_FILE_NAME,
+            flow_file=self.working_dir / flow_file,
             working_dir=self.working_dir,
             dump=False,
         )